[SWBIT-2222] Fixed see development boards
