+++ app/api/hardware/hardware.functions.js
@@ -3,6 +3,7 @@
 var ComponentFunctions = require('./component/component.functions.js'),
     RobotFunctions = require('./robot/robot.functions.js'),
     BoardFunctions = require('./board/board.functions.js'),
+    KitFunctions = require('./kit/kit.functions.js'),
     async = require('async'),
     _ = require('lodash');
 
@@ -17,6 +18,29 @@ exports.getDefault = function(next) {
             boards: result[1],
             components: result[2]
         };
+        next(err, defaultHardware);
+    });
+};
+
+/**
+ * Get hardware
+ */
+exports.getAllHardware = function(next) {
+    async.parallel([
+        RobotFunctions.getAll.bind(RobotFunctions),
+        BoardFunctions.getAll.bind(BoardFunctions),
+        ComponentFunctions.getAll.bind(ComponentFunctions),
+        KitFunctions.getAll.bind(KitFunctions)
+    ], function(err, result) {
+        var defaultHardware = {};
+        if (result) {
+            defaultHardware = {
+                robots: result[0],
+                boards: result[1],
+                components: result[2],
+                kits: result[3]
+            };
+        }
         next(err, defaultHardware);
     });
 };
